feat: changes to STDIN handling for probable perf improvements
I'm not quite sure what improved performance here. But it definitely
seems better. Now, to see any performance issues I need to open a
terminal of size 232x70, open nvim with a large file, and scroll line
by line. It gets sluggish and the cursor does that think were it
dances around the screen. Other than that, I feel like performance
is currently good enough for a beta.

The areas I'd be inclined to look into next for performance improvements
would be the places in which individual terminal cells are parsed:
* The conversion of the virtual PTY to a surface to be stored as shared
  state.
* Compositing tattoys and the PTY surface.

Changes:
* Added `tokio-console` woow.
* Got rid of the `Arc` around the raw PTY file device.
* Removed `std::mem::forget(pty_raw_fd)`. I think all it was doing was
  preventing a second error if the PTY errors. Would be nice to have,
  but since removing the aforementioned `Arc`, it became harder to
  manage.
* Wrapped PTY file device in BufWriter for forwarded STDIN bytes
* Use `portable_pty`'s method for reading PTY outout, rather than the
  PTY's raw file device.
* Upgraded `tokio` to latest.
* I'd still like to really understand the whole thing with the PTY's raw
  file device. Do we need it? Does it have performance implications?
  Does it actually intercept `CTRL+D`?? Maybe we can get rid of it?