feat: refactored tattoys to be more plugin-like
I want the design for internal tattoys to mirror the design for external
tattoys as closely as possible. So I removed access to the shared state
and now I'm only giving tatoys an "input" channel and an "output"
channel. These should resemble the STDIN and STDOUT for plugins.

This new architecture should also help with running tattoys more in
parallel.