fix: always be checking for resize
Previously we were only checking if the end-user's terminal had resized
every time we got a new frame from somewhere. This worked fine when say
a shader was running because that meant that we were checking for
resizes about 30 times per second. However when displaying mostly static
content and then the user resized then the termina wouldn't be
repainted.

I fixed simple by adding a new branch to the main renderer's select loop
whose only job is to check for resizes.

Fixes #72