fix: do not accumlate zero bytes from the PTY
The PTY sends its output over a channel that has messages of a fixed
size of 4096 bytes. Therefore often the end of the message would contain
lots of empty bytes. Usually that wouldn't be a problem but we're also
accumulating PTY output so that we only parse it once the PTY has sent a
batch of messages. That meant that the accumulated bytes were filled
with random swathes of empty data that caused UTF8 parsing to sometimes
break. Instead we know just accumulate non-zero bytes.

Fixes: #11