fix: pixels in bottom of empty cells
Previously a pixel in the lower half of an empty cell would cause the
upper half to be rendered with the default foreground colour, which is
often white, so it just look obviously wrong. The only thing that I had
that was causing this bug was the smoke particles in the smokey cursor
plugin. But obviously more were going to be affected.

Fixes #30