fix: inherit TERM from parent
I think I was mistaken about the need for hardcoding `xterm-256color`.
At the very least we shouldn't be specifying TERM in config when the
convention of inheriting it from the parent is already well established.

It's a whole other kettle of fish whether Tattoy ever gets to the place
where it can act as a polyfill that always runs its Shadow Terminal with
a fixed TERM value.

Fixes #59