fix: support resizing in smokey cursor plugin
Also removed the `find_safe_place()` method. It was supposed to stop
unrealistic popping and explosions and stuff like that. But when I
disabled it it didn't seem to make any difference. And it makes the
cursor look so much cooler when you see the particles emmenating exactly
from its body.

Fixes #3